iris/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-multipart==0.0.9
PyPDF2==3.0.1
httpx==0.27.0
google-generativeai==0.3.2
tqdm==4.66.1
jinja2==3.1.3
rich==13.7.0
anyio==4.3.0
arxiv==1.4.0
python-dotenv==1.0.0
